requirements.txt
crewai
crewai-tools
langchain-community
langchain
langchain-huggingface
langchain-chroma
sentence-transformers
python-dotenv
tavily-python
streamlit
fastapi
uvicorn
